Update 4
Created state file to house state examples.  Created a basic stateful component.  Created a stateful component that assigns a variable within render and displays it to the ui.Created a component that demonstrates updating state by changing a name attribute on button click.  Created a component that updates state and toggles between true/false on button click.  Different ui is displayed depending on the state attribute.  Created a challenges file to house challenge examples.  Created a counter component that has 3 methods which update the state's counter attribute on button click +/- or reset to 0.  The buttons are rendered in JSX as well as the current count as an h1.  Updated the index file to hold the anchor divs for each new jsx file to be rendered onto by ReactDOM.