Update 6
Added example to props file that has a component passing a callback function to a child component as props and rendering.  Added new file to hold life-cycle challenges.  Added component that displays componentWillMount functionality.  Added DidMount challenge that demonstrates using componentDidMount to auto update state after the timeout function completes.  Added challenge that adds an eventlistener to componentDidMount and removes the eventListener with componentWillUnmount.  A message is displayed when the user presses the enter key.  Added challenge that includes componentWillUpdate, componentWillReceiveProps and componentDidUpdate.  It displays a message before the update then compares current props with next props and logs the result.