Update 7
Added example of adding inline style to a react component.  Added component that demonstrates optimizing re-renders with componentShouldUpdate method.  It checks nextProps are even and if they are it updates else it doesnt.  Created a new file to hold rendering examples.  Added a component that demonstrates using conditional operators within the render method to display different ui on button click.  Added a component that does the same but does it by embedding the conditional directly into the JSX and not having to use if/else to display the ui.