Update 8
Added a challenge that a renders an input asking for the user's age.  If no age is entered a submit button is rendered.  If the age is < 18 a button that says You shall not pass is rendered.  Else a button that says you may enter is rendered.  This challenge uses chaining conditional ternary expressions within the JSX code to achieve this.  Added a challenge that creates a 50/50 random roll and passes it value as props to a child component.  The child component returns a win/loss message using a ternary operator that checks the expression is true/false to display a corresponding win/loss statement.