Update 9
Added challenge that changes the border of an input box if the input text length goes beyond 15 characters.  Added a challenge that  takes user text input separated by commas.  It splits the input at the comma into a separate array element.  The array is then mapped into the state and an li JSX element is rendered for each item in the array.