Update 10
added users component to the rendering file.  The component takes an array of user objects and filters them for status online:true.  It then maps the array of online users to an <li> element for each user with the username as text and renders them in a <ul>  Added shouldComponentUpdate method to the game of chance results component.  It tests if the counter has been increased before updating.  Began work stringing the components together by section/file and updating styles and colors for each section.