Update 12
Added challenge example explanations to each component's render method.  Fixed the GameOfChance component's shouldComponentUpdate method to return a boolean instead of undefined.  This was causing it to say you win every time on button click instead of activating the random 50/50 roll and checking it.  This means it wasn't rendering and was just rendering the result of the initial roll each time.